feat: Cleaned up and organized component structure
